pyproject.toml
[project]
name = "pomodoro-bot"
version = "0.1.0"
description = "Voice-driven wake-word + speech-to-text prototype."
requires-python = ">=3.11"
dependencies = [
  "Adafruit-ADS1x15>=1.0.2",
  "adafruit-blinka>=8.0.0",
  "adafruit-circuitpython-ens160>=1.0.8",
  "faster-whisper",
  "google-api-python-client>=2.0.0",
  "google-auth>=2.0.0",
  "huggingface-hub>=0.36.2",
  "llama-cpp-python>=0.3.16",
  "numpy>=1.26.0",
  "piper-tts",
  "psutil>=7.2.2",
  "pvporcupine",
  "pvrecorder",
  "pyinstaller",
  "sounddevice>=0.5.5",
  "transformers<5",
  "typing-extensions",
  "websockets>=15.0.1",
]

[tool.uv]
package = false
